feat(data)!: drop the recorded field bound
MaxDisplacement is gone: the OME-Zarr writer records nothing, the
streamed field write accumulates nothing, and the reader scans nothing
-- the run sizes every region's pull from the field values it reads
for sampling, and the plan prices those reads as a zero field and says
so in its note. What remains of the plan-time scan is the one thing it
still owes: every field entry's header must open, or the group answers
whole-volume before a case is chosen.

BREAKING CHANGE: DISPLACEMENT_BOUND_ATTRIBUTE and displacement_bound
are removed from konfai.utils.ome_zarr; stores carrying the attribute
simply have it ignored.
