+++ docs/source/reference/components/transforms.md
@@ -97,7 +97,7 @@ until it declares otherwise.
 | --- | --- | --- | --- | --- | --- |
 | `Padding` | `F.pad`; updates Origin. `mode` supports `"constant:<val>"`. | `padding=[0,0,0,0,0,0], mode="constant", inverse=True` | **yes** | **yes** | no‡ |
 | `Crop` | Crop to foreground bounding box; caches the box; updates Origin. | `inverse=True` | **yes** | **yes** (pads back) | **yes** — once the `box` is on the case; the region is the patch translated |
-| `Resample` | **The one resample.** Two questions: which grid to write on — nothing (the case's own), `spacing`, `shape`, or `reference` (a stored image's grid, adopted whole) — and what map to write it through — `field` (a displacement field on its own grid, in world units) and/or `transforms` (rigid, affine, BSpline, dense field or composite stored beside the cases; the **last declared is applied first**). Asked for together they compose into **one interpolation**. `align` places a `spacing`/`shape` grid: `extent` keeps the field of view (the outer faces coincide), `origin` keeps voxel zero's centre. `interpolation` left unset is nearest for `uint8` and linear otherwise. | `spacing=None`, `shape=None`, `reference=None`, `reference_group=None`, `reference_dataset=None`, `transforms=None`, `field=None`, `field_group=None`, `align="extent"`, `interpolation=None`, `fill=0.0`, `inverse=True` | **yes** | **yes** — the grid change alone; a declared map is not inverted, and a stage that changes no grid refuses rather than pretend | **yes** — declares `REGRID`. A rigid or affine map bounds exactly; a BSpline and a stored field bound by the sup-norm of their values, which is a theorem (non-negative kernels summing to one), not a sample of the boundary; a field on disk sizes each region's pull from the field values read for sampling anyway — measured at run, per region — and a bound the store recorded at write time prices the plan exactly and is **checked** per component against every region read. Falls back with the reason when the case carries no geometry, a type decomposes into no bounded map, or `invert: true` names a spline or a field |
+| `Resample` | **The one resample.** Two questions: which grid to write on — nothing (the case's own), `spacing`, `shape`, or `reference` (a stored image's grid, adopted whole) — and what map to write it through — `field` (a displacement field on its own grid, in world units) and/or `transforms` (rigid, affine, BSpline, dense field or composite stored beside the cases; the **last declared is applied first**). Asked for together they compose into **one interpolation**. `align` places a `spacing`/`shape` grid: `extent` keeps the field of view (the outer faces coincide), `origin` keeps voxel zero's centre. `interpolation` left unset is nearest for `uint8` and linear otherwise. | `spacing=None`, `shape=None`, `reference=None`, `reference_group=None`, `reference_dataset=None`, `transforms=None`, `field=None`, `field_group=None`, `align="extent"`, `interpolation=None`, `fill=0.0`, `inverse=True` | **yes** | **yes** — the grid change alone; a declared map is not inverted, and a stage that changes no grid refuses rather than pretend | **yes** — declares `REGRID`. A rigid or affine map bounds exactly; a BSpline and a stored field bound by the sup-norm of their values, which is a theorem (non-negative kernels summing to one), not a sample of the boundary; a field on disk sizes each region's pull from the field values read for sampling anyway — measured at run, per region; the plan prices those reads as a zero field, and says so. Falls back with the reason when the case carries no geometry, a type decomposes into no bounded map, or `invert: true` names a spline or a field |
 | `Canonical` | Reorient to canonical direction (3-D); updates Origin/Direction. | `inverse=True` | **yes** — a remap that transposes extents moves the patch grid | **yes** | **yes** — when the case's direction is a signed axis permutation; no on an oblique one (it is resampled) |
 | `Permute` | Permute spatial axes. `dims` is a pipe-separated axis list. | `dims="1\|0\|2", inverse=True` | **yes** | **yes** | **yes** — index remap |
 | `Flip` | Flip spatial axes. | `dims="1\|0\|2", inverse=True` | no | **yes** (self-inverse) | **yes** — index remap |
