+++ tests/unit/test_resample_to_reference.py
@@ -38,7 +38,6 @@ from konfai.data.sampling import source_window
 from konfai.data.transform import LocalityKind, Reduce, Resample, Write
 from konfai.utils.dataset import Attribute, Dataset
 from konfai.utils.errors import ConfigError, TransformError
-from konfai.utils.ome_zarr import DISPLACEMENT_BOUND_ATTRIBUTE
 
 pytest.importorskip("SimpleITK")
 import SimpleITK as sitk
@@ -858,30 +857,9 @@ def test_the_warped_run_never_assembles_the_volume(warped: tuple[Dataset, Datase
     assert list(written.shape[1:]) == list(_REFERENCE_SPATIAL)
 
 
-def test_a_field_beyond_its_recorded_bound_is_refused(
-    warped: tuple[Dataset, Dataset, np.ndarray], tmp_path: Path
-) -> None:
-    """The store's bound is a promise about what was read; data that break it must not sample zeros.
-
-    Sampling past the region that was read gives a dark rim around the moved anatomy and nothing
-    else to see, which is the shape of a mistake nobody finds.
-    """
-    images, _fields, volume = warped
-    attributes = _attributes(_FIELD_ORIGIN, _FIELD_SPACING)
-    attributes[DISPLACEMENT_BOUND_ATTRIBUTE] = np.asarray([0.5, 0.5, 0.5])
-    lying = Dataset(tmp_path / "lying", "mha")
-    lying.write("DVF", _CASE, _displacement(), attributes)
-    stage = Resample(
-        reference=_CASE, reference_group="Reference", field=f"{tmp_path / 'lying'}:mha", field_group="DVF", fill=_FILL
-    )
-    stage.set_datasets([images])
-    with pytest.raises(TransformError, match="displaces up to"):
-        stage(_CASE, torch.from_numpy(volume.copy()), _attributes(_SOURCE_ORIGIN, _SOURCE_SPACING))
-
-
 def test_a_field_with_no_bound_still_streams(warped: tuple[Dataset, Dataset, np.ndarray]) -> None:
     """The run sizes each region's pull from the field
-    values it reads for sampling, so a missing recorded bound is a pricing gap, not a fallback."""
+    values it reads for sampling: the plan prices the field as zero -- a pricing gap, not a fallback."""
     images, fields, _volume = warped
     stage = _warping(images, fields)
     assert stage.patch_locality(_attributes(_SOURCE_ORIGIN, _SOURCE_SPACING)).kind is LocalityKind.REGRID
@@ -936,27 +914,6 @@ def test_fields_can_live_beside_the_cases(warped: tuple[Dataset, Dataset, np.nda
         _CASE, torch.from_numpy(volume.copy()), _attributes(_SOURCE_ORIGIN, _SOURCE_SPACING)
     )
     np.testing.assert_array_equal(got.numpy(), want.numpy())
-
-
-def test_the_recorded_bound_is_read_from_every_root_not_the_first(tmp_path: Path) -> None:
-    """The bound is the cohort's, and a cohort declared by group alone can span the run's roots.
-
-    Stopping at the first root that answers gives a halo sized for part of the cohort: the cases
-    living in the other stores are then read through a region too small for their own displacement.
-    """
-    first, second = Dataset(tmp_path / "first", "mha"), Dataset(tmp_path / "second", "mha")
-    for root, shift in ((first, 1.0), (second, 9.0)):
-        field = np.zeros((3, 4, 4, 4), dtype=np.float32)
-        field[:] = shift
-        attributes = _attributes(_FIELD_ORIGIN, _FIELD_SPACING)
-        attributes[DISPLACEMENT_BOUND_ATTRIBUTE] = np.array([shift, shift, shift])
-        root.write("DVF", f"CASE_{int(shift)}", field, attributes)
-
-    stage = Resample(reference="CASE_1", reference_group="Reference", field_group="DVF")
-    stage.set_datasets([first, second])
-
-    # 9.0 from the second root, not 1.0 from the first.
-    assert stage.displacement.component_bound() == [9.0, 9.0, 9.0]
 
 
 def test_the_two_gathers_obey_the_same_rules_through_an_identity_field(tmp_path: Path) -> None:
